+++ app/src/basic/java/ru/vodolatskii/movies/presentation/fragments/FavoriteFragment.kt
@@ -1,0 +1,30 @@
+package ru.vodolatskii.movies.presentation.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import ru.vodolatskii.movies.databinding.FragmentFavoriteBinding
+
+
+class FavoriteFragment : Fragment() {
+
+    private lateinit var binding: FragmentFavoriteBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentFavoriteBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
