+++ app/src/basic/java/ru/vodolatskii/movies/presentation/fragments/SelectionsFragment.kt
@@ -8,15 +8,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.vodolatskii.movies.databinding.FragmentSelectionsBinding
-import ru.vodolatskii.movies.presentation.utils.AnimationHelper
 
 class SelectionsFragment : Fragment() {
     private lateinit var binding: FragmentSelectionsBinding
-
-
-//    init {
-//        exitTransition = Fade(Fade.MODE_OUT).apply { duration = 500 }
-//    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -26,28 +20,12 @@ class SelectionsFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
+    ): View {
         binding = FragmentSelectionsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-//        AnimationHelper.performFragmentCircularRevealAnimation(view, requireActivity(), 3)
-
-        binding.textedit.addTextChangedListener(object : TextWatcher {
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
-            }
-
-            override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {
-                binding.buttonNext.isEnabled = s.isNotBlank()
-            }
-
-            override fun afterTextChanged(s: Editable?) {
-            }
-
-        })
-
     }
 }
