+++ app/src/main/res/layout/fragment_after.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/after_fragment_root"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/after_fragment_root"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:visibility="invisible"
-    tools:visibility="visible"
     android:background="@drawable/fragments_background"
-    tools:context=".presentation.fragments.AfterFragment">
+    android:visibility="invisible"
+    tools:context=".presentation.fragments.AfterFragment"
+    tools:visibility="visible">
 
-    <TextView
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycler_view_after"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        tools:listitem="@layout/recyclerview_movie_item_layout" />
 
 </FrameLayout>
